backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
crewai
astrapy
python-dotenv
weasyprint
Jinja2
beautifulsoup4
requests
litellm
langchain-google-vertexai
